Fix compile flags for FEDORA and clean up some compilation warnings
git-svn-id: https://mydns-ng.svn.sourceforge.net/svnroot/mydns-ng@173 fff143f6-a9c2-4b89-8f93-78cb2c0259b0

+++ mydns/tags/1.2.8/src/mydns/array.c
@@ -71,7 +71,7 @@ array_free(ARRAY *a, int release_contents) {
 void
 array_append(ARRAY *a, void *o) {
 
-  int idx = a->maxidx + 1;
+  size_t idx = a->maxidx + 1;
 
   if (idx >= a->size) {
     a->objects = (void**)REALLOCATE(a->objects, (a->size*2)*sizeof(void*), void*);
